extracting some address specific functionality to an address class with associated spec - ultimately would be nice if the geocoding was associated with addresses rather than organizations
