+++ app/models/organization.rb
@@ -61,18 +61,7 @@ class Organization < ActiveRecord::Base
     unfriendly_description && unfriendly_description.humanize
   end
 
-  def self.parse_address(address_with_trailing_postcode)  
-    address = address_with_trailing_postcode.to_s
-    postcode = ''
-    match = address_with_trailing_postcode.to_s.match(/(.*)(,\s*(\w\w\d\s* \d\w\w))/)
-    if match
-      if match.length == 4 
-        address = match[1]
-        postcode = match[3].to_s
-      end
-    end
-    {:address => address, :postcode => postcode}
-  end
+
 
   #Edit this if CSV 'schema' changes
   #value is the name of a column in csv file
@@ -122,7 +111,7 @@ class Organization < ActiveRecord::Base
   def self.create_from_array(row, validate)
     check_columns_in(row)
     return nil if row[@@column_mappings[:date_removed]]
-    address = self.parse_address(row[@@column_mappings[:address]])
+    address = Address.parse_address(row[@@column_mappings[:address]])
 
     org = Organization.new
     org.name = row[@@column_mappings[:name]].to_s.humanized_all_first_capitals
